+++ examples/resources/buildkite_registry/resource.tf
@@ -1,0 +1,7 @@
+resource "buildkite_registry" "example" {
+  name = "example"
+  description = "super cool ruby registry"
+  ecosystem = "ruby"
+  emoji = ":ruby:"
+  color = "#ff0000"
+}